refactor: rename pipeline/ → primitives/
Aligns folder naming with the render lifecycle (bottom-up):
  primitives/  ← 0. low-level ops (ffmpeg, Groq API)
  renderers/   ← 1. per-element generation
  ir/          ← 2. compile → execute (planning)
  render.ts    ← 3. top-level orchestrator

'pipeline' described a programming pattern, not a domain or layer.
'primitives' says what these modules are: the lowest-level operations
that renderers build upon.
